chore: restructure repo
