+++ build.sh
@@ -2,5 +2,5 @@ export PATH="$PATH:$HOME/.nargo/bin/"
 curl -L https://raw.githubusercontent.com/noir-lang/noirup/main/install | bash
 noirup -n
 nargo -V
-cd src/noir-script
+cd noir-script
 nargo compile noir-script
