--- my-element.js
@@ -1,3 +1,0 @@
-import { MyElement } from "./src/modules/my_element";
-
-window.customElements.define("my-element", MyElement);
