+++ package.json
@@ -20,6 +20,7 @@
     "@web/dev-server-esbuild": "^0.3.6",
     "@web/test-runner": "^0.15.3",
     "@web/test-runner-playwright": "^0.10.0",
+    "chai": "^4.3.7",
     "lit": "^2.7.2",
     "mocha": "^10.2.0",
     "ts-node": "^10.9.1",
