+++ web-test-runner.config.mjs
@@ -5,16 +5,14 @@ export default {
   browsers: [
     playwrightLauncher({ product: "chromium" }),
     playwrightLauncher({ product: "webkit" }),
-    // Firefox requires 40s to perform a Pedersen hash so we recommend using either
-    // a Chromium- or Webkit-based browser
-    // playwrightLauncher({ product: "firefox" }),
+    playwrightLauncher({ product: "firefox" }),
   ],
   plugins: [
     esbuildPlugin({
       ts: true,
     }),
   ],
-  files: ["src/test/**/*.test.ts"],
+  files: ["test/browser/**/*.test.ts"],
   nodeResolve: true,
   testFramework: {
     config: {
